feat(UNS-105+104): no-JS static-HTML edge function for /a/*, remove old claimed-artist-page
- New: api/edge/artist-page-static.ts — serves fully-rendered static HTML
  for both claimed and unclaimed artists at /a/* with ZERO JavaScript
- Deleted: api/edge/claimed-artist-page.ts (452 lines, replaced by
  artist-page-static.ts in this same PR)
- Modified: netlify.toml — swap /a/* edge function from
  claimed-artist-page to artist-page-static
- Modified: apps/web/vite.config.ts — remove /artist/* from
  navigateFallbackDenylist (old redirect target no longer used)
- Modified: CLAUDE.md — update edge functions list

Key design decisions:
- No JS in response: theme via CSS prefers-color-scheme, no auth UI,
  no analytics tracking, no embed widget JS
- XSS safety: escapeHtml() on every user-supplied string; featured embed
  iframe restricted to ALLOWED_EMBED_DOMAINS allowlist with
  sandbox="allow-scripts allow-same-origin"
- 5-second AbortController timeout on Supabase fetches; falls through
  to SPA on timeout/error
- Claimed artists: full profile with bio, featured embed, Verified badge,
  platform links with payout %
- Unclaimed artists: simpler card with platform links, no bio/embed/badge,
  plus Claim CTA linking to /claim?slug={slug}
- JSON-LD structured data for both claimed and unclaimed (MusicGroup);
  unclaimed omits speculative sameAs

Reviewers: @dan (XSS + no-JS verification)
