+++ apps/web/vite.config.ts
@@ -70,7 +70,6 @@ export default defineConfig({
         navigateFallbackDenylist: [
           // Edge function routes that serve static HTML — don't let SW intercept them
           /^\/a\//,
-          /^\/artist\//,
         ],
         runtimeCaching: [
           {
